Move DTOs related to Authentication to separate folders 'Login', 'Register' and 'RefreshToken'
