+++ Streetcode/Streetcode.BLL/DTO/Authentication/RefreshToken/RefreshTokenRequestDTO.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Streetcode.BLL.DTO.Users
+namespace Streetcode.BLL.DTO.Authentication.RefreshToken
 {
-    public class RefreshTokenDTO
+    public class RefreshTokenRequestDTO
     {
         public string Token { get; set; }
     }
