--- Streetcode/Streetcode.BLL/DTO/Users/RefreshTokenResponce.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Streetcode.BLL.DTO.Users
-{
-    public class RefreshTokenResponce
-    {
-        public string Token { get; set; }
-        public DateTime ExpireAt { get; set; }
-    }
-}
