+++ Streetcode/Streetcode.BLL/Mapping/Authentication/AuthProfile.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using Streetcode.BLL.DTO.Authentication;
+using Streetcode.BLL.DTO.Authentication.Register;
 using Streetcode.BLL.DTO.Users;
 using Streetcode.DAL.Entities.Users;
