+++ Streetcode/Streetcode.BLL/MediatR/Authentication/Login/LoginQuery.cs
@@ -1,6 +1,6 @@
 ﻿using FluentResults;
 using MediatR;
-using Streetcode.BLL.DTO.Authentication;
+using Streetcode.BLL.DTO.Authentication.Login;
 using Streetcode.BLL.DTO.Users;
 
 namespace Streetcode.BLL.MediatR.Users.Login
