+++ Streetcode/Streetcode.BLL/MediatR/Authentication/RefreshToken/RefreshTokenHandler.cs
@@ -2,7 +2,7 @@
 using AutoMapper;
 using FluentResults;
 using MediatR;
-using Streetcode.BLL.DTO.Users;
+using Streetcode.BLL.DTO.Authentication.RefreshToken;
 using Streetcode.BLL.Interfaces.Logging;
 using Streetcode.BLL.Interfaces.Users;
 using Streetcode.BLL.Services.Users;
@@ -10,7 +10,7 @@ using Streetcode.DAL.Repositories.Interfaces.Base;
 
 namespace Streetcode.BLL.MediatR.Users.RefreshToken
 {
-    public class RefreshTokenHandler : IRequestHandler<RefreshTokenQuery, Result<RefreshTokenResponce>>
+    public class RefreshTokenHandler : IRequestHandler<RefreshTokenQuery, Result<RefreshTokenResponceDTO>>
     {
         private readonly IMapper _mapper;
         private readonly IRepositoryWrapper _repositoryWrapper;
@@ -23,7 +23,7 @@ namespace Streetcode.BLL.MediatR.Users.RefreshToken
             _tokenService = tokenService;
         }
 
-        public async Task<Result<RefreshTokenResponce>> Handle(RefreshTokenQuery request, CancellationToken cancellationToken)
+        public async Task<Result<RefreshTokenResponceDTO>> Handle(RefreshTokenQuery request, CancellationToken cancellationToken)
         {
             JwtSecurityToken? token = null;
             await Task.Run(() =>
@@ -31,7 +31,7 @@ namespace Streetcode.BLL.MediatR.Users.RefreshToken
                 token = _tokenService.RefreshToken(request.token.Token);
             });
 
-            return new RefreshTokenResponce() { Token = new JwtSecurityTokenHandler().WriteToken(token), ExpireAt = token !.ValidTo };
+            return new RefreshTokenResponceDTO() { Token = new JwtSecurityTokenHandler().WriteToken(token), ExpireAt = token !.ValidTo };
         }
     }
 }
