+++ Streetcode/Streetcode.BLL/MediatR/Authentication/RefreshToken/RefreshTokenQuery.cs
@@ -1,8 +1,8 @@
 ﻿using FluentResults;
 using MediatR;
-using Streetcode.BLL.DTO.Users;
+using Streetcode.BLL.DTO.Authentication.RefreshToken;
 
 namespace Streetcode.BLL.MediatR.Users.RefreshToken
 {
-    public record RefreshTokenQuery(RefreshTokenDTO token) : IRequest<Result<RefreshTokenResponce>>;
+    public record RefreshTokenQuery(RefreshTokenRequestDTO token) : IRequest<Result<RefreshTokenResponceDTO>>;
 }
