+++ Streetcode/Streetcode.BLL/MediatR/Authentication/Register/RegisterHandler.cs
@@ -4,7 +4,7 @@ using FluentResults;
 using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
-using Streetcode.BLL.DTO.Authentication;
+using Streetcode.BLL.DTO.Authentication.Register;
 using Streetcode.BLL.DTO.Users;
 using Streetcode.BLL.Interfaces.Logging;
 using Streetcode.DAL.Entities.Users;
