+++ Streetcode/Streetcode.WebApi/Controllers/Authentication/AuthController.cs
@@ -1,8 +1,9 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Streetcode.BLL.DTO.Authentication;
-using Streetcode.BLL.DTO.Users;
+using Streetcode.BLL.DTO.Authentication.Login;
+using Streetcode.BLL.DTO.Authentication.RefreshToken;
+using Streetcode.BLL.DTO.Authentication.Register;
 using Streetcode.BLL.MediatR.Users.Login;
 using Streetcode.BLL.MediatR.Users.RefreshToken;
 using Streetcode.BLL.MediatR.Users.SignUp;
@@ -36,7 +37,7 @@ namespace Streetcode.WebApi.Controllers.Users
         }
 
         [HttpPost]
-        public async Task<IActionResult> RefreshToken([FromBody] RefreshTokenDTO token)
+        public async Task<IActionResult> RefreshToken([FromBody] RefreshTokenRequestDTO token)
         {
             return HandleResult(await Mediator.Send(new RefreshTokenQuery(token)));
         }
